+++ rds-20140815/src/Models/RunRCInstancesShrinkRequest.php
@@ -37,6 +37,11 @@ class RunRCInstancesShrinkRequest extends Model
     /**
      * @var string
      */
+    public $businessInfo;
+
+    /**
+     * @var string
+     */
     public $clientToken;
 
     /**
@@ -234,6 +239,7 @@ class RunRCInstancesShrinkRequest extends Model
         'autoPay' => 'AutoPay',
         'autoRenew' => 'AutoRenew',
         'autoUseCoupon' => 'AutoUseCoupon',
+        'businessInfo' => 'BusinessInfo',
         'clientToken' => 'ClientToken',
         'createAckEdgeParamShrink' => 'CreateAckEdgeParam',
         'createExtraParam' => 'CreateExtraParam',
@@ -304,6 +310,10 @@ class RunRCInstancesShrinkRequest extends Model
 
         if (null !== $this->autoUseCoupon) {
             $res['AutoUseCoupon'] = $this->autoUseCoupon;
+        }
+
+        if (null !== $this->businessInfo) {
+            $res['BusinessInfo'] = $this->businessInfo;
         }
 
         if (null !== $this->clientToken) {
@@ -498,6 +508,10 @@ class RunRCInstancesShrinkRequest extends Model
 
         if (isset($map['AutoUseCoupon'])) {
             $model->autoUseCoupon = $map['AutoUseCoupon'];
+        }
+
+        if (isset($map['BusinessInfo'])) {
+            $model->businessInfo = $map['BusinessInfo'];
         }
 
         if (isset($map['ClientToken'])) {